Camera: Modify for board change
Because s5k3m2 need 1.05V DVDD, so separate it to front camera.
Then in board with s5k3m2, front camera use gpio 117 to control
DVDD. Rear camera use gpio 118 to control DVDD.
In 8M sensor, we also use 1.2V to supply DVDD. So we use old
board which gpio 118 control DVDD for both front and rear camera.
To compatibily with the two board, we distribut gpio 117 and 118
to front camera.

Signed-off-by: liyatang <liyatang@longcheer.net>